Module core_path::channel_graph

Structs

ChannelEdge

Structure that adds additional data to a ChannelEntry, which can be used to compute edge weights and traverse the ChannelGraph.

ChannelGraph

Implements a HOPR payment channel graph (directed) cached in-memory. This structure is useful for tracking channel state changes and packet path finding. The structure is updated only from the Indexer and therefore contains only the channels that were seen on-chain. The network qualities are also added to the graph on the fly. Using this structure is much faster than querying the DB and therefore is preferred for per-packet path-finding computations. Per default, the graph does not track channels in Closed state, and therefore cannot detect channel re-openings.