core_path::selectors::legacy
Type Alias LegacyPathSelector

pub type LegacyPathSelector = DfsPathSelector<RandomizedEdgeWeighting>;

Legacy DFS path selector with channel weighting function that uses randomized channel stakes as edge weights.

Aliased Type

struct LegacyPathSelector {
    pub max_iterations: usize,
    pub quality_threshold: f64,
    pub allow_zero_edge_weight: bool,
    /* private fields */
}

Fields
max_iterations: usize

Maximum number of iterations before a path selection fails Default is 100

quality_threshold: f64

Peer quality threshold for a channel to be taken into account. Default is 0.5

allow_zero_edge_weight: bool

If true, include paths with payment channels, which have no funds in it. By default, that option is set to false to prevent from tickets being dropped immediately. Defaults to false.