Type Alias core_path::selectors::legacy::LegacyPathSelector

pub type LegacyPathSelector = DfsPathSelector<RandomizedEdgeWeighting>;

Legacy DFS path selector with channel weighting function that uses randomized channel stakes as edge weights.

Aliased Type

struct LegacyPathSelector {
    pub max_iterations: usize,
    pub quality_threshold: f64,
    /* private fields */
}

Fields
max_iterations: usize

Maximum number of iterations before a path selection fails Default is 100

quality_threshold: f64

Peer quality threshold for a channel to be taken into account. Default is 0.5